Handle intermediate boots without SELinux
If a system had SELinux on, then the Admin boots with SELinux disabled
(instead of permissive), changes some files and boots with SELinux
enabled again, we may end up with total wreckage and unbooting system
due to unlabelled files.
Fortunately if a system was ever labelled due to SELinux we can find the
/etc/selinux/.relabelled marker. If that file is present but SELinux off
we know it was turned off again. So touch /etc/selinux/.autorelabel to
make the system relabels itself in case SELinux gets turned on again.