Fix unwritable /var / /etc after SELinux relabel
mount -o remount,ro will propagate the read-only attribute also to subvolumes
on the same partition, effectively making /var and /etc unwritable if they are
located on the same device as it is the case with some of our images. Use
an alternative mountpoint instead to preserve the permissions of the original
mounts.
[bsc#1186563]